requirements.txt
Django==4.2.6
django-filter==23.3
djangorestframework==3.14.0
django-crispy-forms==2.1
django-bootstrap-v5==1.0.11
crispy-bootstrap5==2023.10
Faker==19.12.0
psycopg2-binary==2.9.9
gunicorn==21.2.0
django-cors-headers==4.6.0